requirements.txt
absl-py==0.15.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==4.2.4
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
click==8.1.2
cssselect2==0.5.0
cycler==0.11.0
defusedxml==0.7.1
fastapi==0.75.1
filelock==3.6.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py==3.1.0
html5lib==1.1
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
jarowinkler==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.4.2
mplcursors==0.5.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pypdfium2==1.1.0
pyphen==0.12.0
python-dateutil==2.8.2
python-doctr==0.5.1
python-multipart==0.0.5
PyYAML==6.0
rapidfuzz==2.0.7
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
Shapely==1.8.1.post1
six==1.15.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.3
tensorflow-addons==0.16.1
tensorflow-estimator==2.6.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tinycss2==1.1.1
tqdm==4.63.0
typeguard==2.13.3
typing-extensions==3.10.0.2
Unidecode==1.3.4
urllib3==1.26.9
uvicorn==0.17.6
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0
